requirements.txt
requests
beautifulsoup4
python-dateutil
tzdata
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
